Refactored generate_helper_bits into a function
