+++ src/protocol/attribution/malicious.rs
@@ -2,23 +2,22 @@ use super::{
     accumulate_credit::accumulate_credit,
     aggregate_credit::malicious_aggregate_credit,
     credit_capping::credit_capping,
+    generate_helper_bits,
     input::{MCAccumulateCreditInputRow, MCAggregateCreditOutputRow},
 };
 use crate::{
     error::Error,
     ff::{GaloisField, PrimeField, Serializable},
     protocol::{
-        boolean::{equality_test, random_bits_generator::RandomBitsGenerator},
         context::{Context, SemiHonestContext},
         ipa::IPAModulusConvertedInputRow,
         malicious::MaliciousValidator,
-        RecordId, Substep,
+        Substep,
     },
     secret_sharing::replicated::{
         malicious::AdditiveShare, semi_honest::AdditiveShare as SemiHonestAdditiveShare,
     },
 };
-use futures::future::try_join_all;
 use std::iter::zip;
 
 /// Performs a set of attribution protocols on the sorted IPA input.
@@ -41,28 +40,7 @@ where
 {
     let m_ctx = malicious_validator.context();
 
-    let eq_test_ctx = m_ctx
-        .narrow(&Step::ComputeHelperBits)
-        .set_total_records(sorted_rows.len() - 1);
-
-    let rbg = RandomBitsGenerator::new(m_ctx.clone());
-
-    let futures = sorted_rows
-        .iter()
-        .zip(sorted_rows.iter().skip(1))
-        .enumerate()
-        .map(|(i, (row, next_row))| {
-            let record_id = RecordId::from(i);
-            let c = eq_test_ctx.clone();
-            let rbg_ref = &rbg;
-            async move {
-                equality_test(c, record_id, rbg_ref, &row.mk_shares, &next_row.mk_shares).await
-            }
-        });
-
-    let helper_bits = Some(AdditiveShare::ZERO)
-        .into_iter()
-        .chain(try_join_all(futures).await?);
+    let helper_bits = generate_helper_bits(m_ctx.clone(), &sorted_rows).await?;
 
     let attribution_input_rows = zip(sorted_rows, helper_bits)
         .map(|(row, hb)| {
@@ -104,7 +82,6 @@ where
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum Step {
-    ComputeHelperBits,
     AccumulateCredit,
     PerformUserCapping,
     AggregateCredit,
@@ -115,7 +92,6 @@ impl Substep for Step {}
 impl AsRef<str> for Step {
     fn as_ref(&self) -> &str {
         match self {
-            Self::ComputeHelperBits => "compute_helper_bits",
             Self::AccumulateCredit => "accumulate_credit",
             Self::PerformUserCapping => "user_capping",
             Self::AggregateCredit => "aggregate_credit",
